+++ Chapter 13 MVC/controllers/404NoutFound.js
@@ -1,0 +1,4 @@
+
+exports.NotFoundPage = (req, res, next) => {
+  res.status(404).render('404NotFound',{pageTitle:"Page not found.", currentPage:"404NotFound"});;
+};
